requirements.txt
PyYAML~=6.0
empire_commons@https://github.com/Tombmyst-Empire/empire-commons/archive/refs/heads/master.zip